chore: move NavigationBar to components
